requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0
bcrypt==4.0.1
openai==1.3.5
python-google-auth==1.0.0
google-auth-oauthlib==1.1.0
python-jose[cryptography]==3.3.0
gTTS==2.3.2
